+++ toolkit/address-associations/pallet/src/tests.rs
@@ -4,40 +4,32 @@ use frame_system::pallet_prelude::OriginFor;
 use hex_literal::hex;
 use mock::*;
 use sidechain_domain::*;
-use sp_core::crypto::Ss58Codec;
 use sp_runtime::AccountId32;
 
 #[test]
 fn saves_new_address_association() {
 	new_test_ext().execute_with(|| {
-			// Alice
-			let stake_public_key = StakePublicKey(hex!(
-				"2bebcb7fbc74a6e0fd6e00a311698b047b7b659f0e047ff5349dbd984aefc52c"
-			));
-			let mc_signature = hex!("1aa8c1b363a207ddadf0c6242a0632f5a557690a327d0245f9d473b983b3d8e1c95a3dd804cab41123c36ddbcb7137b8261c35d5c8ef04ce9d0f8d5c4b3ca607");
+		assert_eq!(mock_pallet::LastNewAssociation::<Test>::get(), None);
+		let initial_balance = Balances::free_balance(&FUNDED_ACCOUNT);
+		assert_ok!(super::Pallet::<Test>::associate_address(
+			OriginFor::<Test>::signed(FUNDED_ACCOUNT),
+			pc_address(),
+			VALID_SIGNATURE.into(),
+			STAKE_PUBLIC_KEY.clone(),
+		));
 
-			// Alice
-			let pc_address =
-					AccountId32::from_ss58check("5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY").unwrap();
+		assert_eq!(
+			Pallet::<Test>::get_partner_chain_address_for(&STAKE_PUBLIC_KEY),
+			Some(pc_address())
+		);
 
-			assert_eq!(mock_pallet::LastNewAssociation::<Test>::get(), None);
-
-			assert_ok!(
-				super::Pallet::<Test>::associate_address(
-					OriginFor::<Test>::signed(AccountId32::new([1; 32])),
-					pc_address.clone(),
-					mc_signature.into(),
-					stake_public_key.clone(),
-				)
-			);
-
-			assert_eq!(
-				Pallet::<Test>::get_partner_chain_address_for(&stake_public_key),
-				Some(pc_address.clone())
-			);
-
-			assert_eq!(mock_pallet::LastNewAssociation::<Test>::get(), Some((pc_address,stake_public_key.hash())));
-		})
+		assert_eq!(
+			mock_pallet::LastNewAssociation::<Test>::get(),
+			Some((pc_address(), STAKE_PUBLIC_KEY.hash()))
+		);
+		let final_balance = Balances::free_balance(&FUNDED_ACCOUNT);
+		assert_eq!(final_balance, initial_balance - AssociationFeeBurn::get());
+	})
 }
 
 #[test]
@@ -56,7 +48,7 @@ fn rejects_duplicate_key_association() {
 		);
 		assert_eq!(
 			super::Pallet::<Test>::associate_address(
-				OriginFor::<Test>::signed(AccountId32::new([1; 32])),
+				OriginFor::<Test>::signed(FUNDED_ACCOUNT),
 				AccountId32::new([0; 32]),
 				signature,
 				stake_public_key,
@@ -70,20 +62,34 @@ fn rejects_duplicate_key_association() {
 #[test]
 fn rejects_invalid_mainchain_signature() {
 	new_test_ext().execute_with(|| {
-			let stake_public_key = StakePublicKey(hex!(
-				"fc014cb5f071f5d6a36cb5a7e5f168c86555989445a23d4abec33d280f71aca4"
-			));
-			let signature = StakeKeySignature(hex!("c50828c31d1a61e05fdb943847efd42ce2eadda9c7d21dd2d035e8de66bc56de7f6b1297fba6cb7305f2aac97b5f9168894fb10295c503de6d5fb6ae70bd9a0d"));
+
+			let invalid_signature = StakeKeySignature(hex!("c50828c31d1a61e05fdb943847efd42ce2eadda9c7d21dd2d035e8de66bc56de7f6b1297fba6cb7305f2aac97b5f9168894fb10295c503de6d5fb6ae70bd9a0d"));
 
 			assert_eq!(
 				super::Pallet::<Test>::associate_address(
-					OriginFor::<Test>::signed(AccountId32::new([1; 32])),
-					AccountId32::new([0; 32]),
-					signature,
-					stake_public_key,
+					OriginFor::<Test>::signed(FUNDED_ACCOUNT),
+					pc_address(),
+					invalid_signature,
+					STAKE_PUBLIC_KEY,
 				)
 				.unwrap_err(),
 				Error::<Test>::InvalidMainchainSignature.into()
 			);
 		})
+}
+
+#[test]
+fn rejects_extrinsic_when_origin_account_cannot_pay_extra_fee() {
+	new_test_ext().execute_with(|| {
+		assert_eq!(
+			super::Pallet::<Test>::associate_address(
+				OriginFor::<Test>::signed(AccountId32::new([3; 32])),
+				pc_address(),
+				VALID_SIGNATURE.into(),
+				STAKE_PUBLIC_KEY,
+			)
+			.unwrap_err(),
+			Error::<Test>::InsufficientBalance.into()
+		);
+	})
 }