+++ demo/runtime/src/lib.rs
@@ -367,7 +367,7 @@ impl pallet_balances::Config for Runtime {
 	type WeightInfo = pallet_balances::weights::SubstrateWeight<Runtime>;
 	type FreezeIdentifier = ();
 	type MaxFreezes = ();
-	type RuntimeHoldReason = ();
+	type RuntimeHoldReason = RuntimeHoldReason;
 	type RuntimeFreezeReason = RuntimeFreezeReason;
 	type DoneSlashHandler = ();
 }
@@ -554,10 +554,17 @@ impl pallet_block_production_log::Config for Runtime {
 	type BenchmarkHelper = PalletBlockProductionLogBenchmarkHelper;
 }
 
+parameter_types! {
+	/// Amount of tokens to burn when making irreversible, forever association
+	pub const AddressAssociationBurnAmount: Balance = 1_000_000;
+}
+
 impl pallet_address_associations::Config for Runtime {
 	type WeightInfo = pallet_address_associations::weights::SubstrateWeight<Runtime>;
 
 	type PartnerChainAddress = AccountId;
+	type Currency = Balances;
+	type BurnAmount = AddressAssociationBurnAmount;
 
 	fn genesis_utxo() -> UtxoId {
 		Sidechain::genesis_utxo()
@@ -592,6 +599,11 @@ impl pallet_block_producer_fees::Config for Runtime {
 	type BenchmarkHelper = PalletBlockProducerFeesBenchmarkHelper;
 }
 
+parameter_types! {
+	/// Amount of tokens to hold when upserting block producer metadata.
+	pub const MetadataHoldAmount: Balance = 1_000_000;
+}
+
 impl pallet_block_producer_metadata::Config for Runtime {
 	type WeightInfo = pallet_block_producer_metadata::weights::SubstrateWeight<Runtime>;
 
@@ -600,6 +612,10 @@ impl pallet_block_producer_metadata::Config for Runtime {
 	fn genesis_utxo() -> UtxoId {
 		Sidechain::genesis_utxo()
 	}
+
+	type Currency = Balances;
+	type HoldAmount = MetadataHoldAmount;
+	type RuntimeHoldReason = RuntimeHoldReason;
 
 	#[cfg(feature = "runtime-benchmarks")]
 	type BenchmarkHelper = PalletBlockProducerMetadataBenchmarkHelper;